Added sensitive files, updated processed files in real-time, and separated concerns by adding utility functions and process logic to utils.py
